// src/app/shared/generated/model/configuration.ts
/**
 * onecx-ai bff
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MCPServer } from './mCPServer';
import { ConfigurationFilter } from './configurationFilter';
import { Provider } from './provider';


export interface Configuration { 
    modificationCount?: number;
    creationDate?: string;
    creationUser?: string;
    modificationDate?: string;
    modificationUser?: string;
    id?: string;
    name: string;
    description?: string;
    filter?: ConfigurationFilter;
    llmSystemMessage?: string;
    llmProvider?: Provider;
    mcpServers?: Array<MCPServer>;
}